pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "podcast-llm-wiki"
version = "0.1.0"
description = "Ingest podcast playlists, transcribe locally, and compound analyses into a Karpathy-style LLM Wiki in Obsidian."
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.11"
authors = [{ name = "Geoffrey Byers" }]
dependencies = [
    "pydantic>=2.6,<3",
    "pyyaml>=6.0",
    "typer>=0.12,<1",
    "yt-dlp>=2024.1.0",
    "faster-whisper>=1.0",
    "python-dotenv>=1.0",
    "pyannote.audio>=3.1",
    "torch>=2.1",  # required by pyannote
    "soundfile>=0.12",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-cov>=4.1",
]

[project.scripts]
podcast-llm-wiki = "podcast_llm_wiki.cli:app"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
markers = [
    "integration: real-pipeline smoke tests; require fixture audio and network. Skipped unless explicitly selected.",
]
addopts = "-m 'not integration'"
